requirements.txt
azure_identity==1.17.1
openai==1.57.2
pandas==2.2.2
Requests==2.32.3
tqdm==4.66.2
vllm